+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - New XP formula.
 - The way some cooldowns work.
 - $give/transfer maximum transfer amount increased to $5,000,000.
+- $weekly now gives you a random amount of money and has been nerfed, but scales with prestige now.
 
 ### Fixed
 - Possible sources of data loss.